Fix the file open string
File -> File don't make sense and in not consistent with nearly all other™
Software around (Inkscape, Gimp, LO, Geany and so on)

Should be Open… (with Ellipsis instead of three point). We should change
'...' to '…' projectwide, all people will hate us and esp. me, esp the translators,
but hey - we really should do that.