Prevent blurry images with HDPI
Fixes https://github.com/lxqt/lximage-qt/issues/242

Previously, when Qt scale factor was greater than 1, (unscaled) images became blurry. This patch fixes the problem by setting the device pixel ratio of the image to that of the app and also by considering it in calculating sizes. In short, images are shown with their real sizes.

Also cleaned up the code for image outline and fixed a small issue in it.