Added "Open With.." to File menu
It will be enabled only if a file is opened in lximage-qt.

NOTE: The code could be enhanced so that pasted images or unsaved screenshots would also be opened in other apps. That would require saving them in `/tmp` but I don't think it's a good idea because the temporary file would remain there after being opened by another app. That's the case with apps like Screengrab too (when Screengrab exits *after* its unsaved screenshot is opened by another app).

Closes https://github.com/lxqt/lximage-qt/issues/289