Fixed maximization at startup with SVG and GIF images
The root of this minor problem was deeper than meets the eye: with SVG and GIF, the maximization info was lost when we resized the window before showing it. This may be a bug in Qt. The patch remembers the maximization info and applies it when the window is going to be shown.

Fixes https://github.com/lxqt/lximage-qt/issues/301