fix bug #307 : drawing rectangle doesn't show
The reason for this bug is that when paintdevice is a widget,
QPainter can only be used inside a paintEvent() function or in a
function called by paintEvent()

this commit fix this bug by drawing the same rectangle in the scene

LXImage use QGraphicsView to display image and QGraphicsView is a
part of the Grapics View Framework. Usually by calling update(), a
widget is able to trigger paintEvent(). However, as far as i know,
in the Graphics View Framework, it's impossible for QGraphicsView
to tigger the paintEvent() by calling update()

so i choose the Graphics View Framwork way to draw the rectangle
something like: scene.addRect(...)