fix bug #307 : drawing rectangle doesn't show (#319)
* fix bug #307 : drawing rectangle doesn't show

The reason for this bug is that when paintdevice is a widget,
QPainter can only be used inside a paintEvent() function or in a
function called by paintEvent()

this commit fix this bug by drawing the same rectangle in the scene

LXImage use QGraphicsView to display image and QGraphicsView is a
part of the Grapics View Framework. Usually by calling update(), a
widget is able to trigger paintEvent(). However, as far as i know,
in the Graphics View Framework, it's impossible for QGraphicsView
to tigger the paintEvent() by calling update()

so i choose the Graphics View Framwork way to draw the rectangle
something like: scene.addRect(...)

* fix annatation persistence problem

remove all annotaion items when reloading a new image or rotating
the current image

remove all annotaion items in the scene when reloading image

* Fix the possible memory leak and crash

In setImage() function, when the image is gif or svg, scene_->clear()
is called, which clear and delete the annotation items, so we just
need clear the annotations list, delete or remove the annotations
will make a crash
on the other hand, for a normal image, scene_->removeItem() just
remove the item but not delete it, so we need delete it by ourself
to prevent a memory leak
At last, we need clear the annotations list for both situations

* replace QGraphicsTextItem with QGraphicsSimpleTextItem

avoid magic number and put the text in the center

* remove the const modifier for drawArrow(...)

use the private member directly and make it more reasonable