Show error message if saving with a format is not supported
For example, images cannot be saved as SVG or GIF and it's better to inform the user about that, instead of creating an empty file or, worse, overwrite an existing image.

Also, corrected `MainWindow::getImageGraphicsItem()` because it didn't take a recent annotations commit into account (annotations are now added as items).