Prevent users from defining ambiguous shortcuts (#331)
* Prevent users from defining ambiguous shortcuts

If the user tries to define an ambiguous shortcut, the shortcut will not change but a temporary warning bar will appear at the bottom of Preferences dialog for 2.5 seconds.

In addition, if an ambiguous shortcut was defined by editing the config file (while lximage-qt was closed) or by an older version of lximage-qt, a permanent warning bar will appear. It will be removed once the ambiguity is resolved.

Also:

 * Added a tooltip about shortcut removal; and
 * Fixed a bug in defining a shortcut for the annotation toolbar.

NOTE: I'll update the translations after this is merged.

Closes https://github.com/lxqt/lximage-qt/issues/330

* Also, close warning bar on restoring shortcuts

In rare cases on manual editing.