Fixed invalid nullptr parameter in `QObject::connect`
`QNetworkAccessManager` is changed to a static `QPointer`, which is deleted when all instances of `Provider` are deleted.

In addition, all providers are created on the heap, as children of the upload dialog.

In this way, the following warning is silenced:

```
qt.core.qobject.connect: QObject::connect(QObject, Unknown): invalid nullptr parameter
```