Supported multi-page TIFF images
If a multi-page TIFF image is opened, the menu-items "Next Frame" and "Previous Frame" will be added to the "Go" menu, with the shortcuts "Ctrl+Right" and "Ctrl+Left" respectively. Of course the shortcuts can be changed in the shortcut editor.

To prevent cluttering of the GUI for this special use case, I didn't add another way of informing the users. They should either check the "Go" menu or press those shortcuts to see whether a TIFF image includes multiple pages.

Closes https://github.com/lxqt/lximage-qt/issues/721